requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
blessed==1.20.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorlog==6.7.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography @ file:///croot/cryptography_1689373676338/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.9.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.57.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1698244021190/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081186/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lightning-utilities==0.9.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1693812469450/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
munkres==1.1.4
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx @ file:///croot/networkx_1690561992265/work
numexpr @ file:///croot/numexpr_1683221822650/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-ml-py==12.535.108
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.1
pandas @ file:///croot/pandas_1692289311655/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
ply==3.11
pooch @ file:///croot/pooch_1695850093751/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.24.2
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
pyrootutils==1.0.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytorch-lightning==2.0.8
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scipy==1.10.1
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///croot/sympy_1668202399572/work
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.1.1
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1698232018516/work
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
zipp @ file:///croot/zipp_1672387121353/work
